+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Export more functions to help constructing skins
 * trField to translate field names
 * Allow processOptions to manage with standard control props
+* Allow button and form skinage
 
 ### Fixed